+++ src/vctrs.h
@@ -383,10 +383,11 @@ SEXP vec_recycle_fallback(SEXP x, R_len_t size, struct vctrs_arg* x_arg);
 SEXP vec_recycle_common(SEXP xs, R_len_t size);
 SEXP vec_names(SEXP x);
 SEXP vec_group_loc(SEXP x);
-SEXP vec_match_params(SEXP needles, SEXP haystack, bool na_equal);
+SEXP vec_match_params(SEXP needles, SEXP haystack, bool na_equal,
+                      struct vctrs_arg* needles_arg, struct vctrs_arg* haystack_arg);
 
 static inline SEXP vec_match(SEXP needles, SEXP haystack) {
-  return vec_match_params(needles, haystack, true);
+  return vec_match_params(needles, haystack, true, NULL, NULL);
 }